--- components/ui/Answer/Answer.tsx
@@ -1,25 +1,0 @@
-import React, { useEffect, useState } from "react";
-
-import styles from "./answer.module.css";
-
-interface AnswerProps {
-  text: string;
-}
-
-export const Answer: React.FC<AnswerProps> = ({ text }) => {
-  const [words, setWords] = useState<string[]>([]);
-
-  useEffect(() => {
-    setWords(text.split(" "));
-  }, [text]);
-
-  return (
-    <div>
-      {words.map((word, index) => (
-        <span key={index} className={styles.fadeIn} style={{ animationDelay: `${index * 0.01}s` }}>
-          {word}{" "}
-        </span>
-      ))}
-    </div>
-  );
-};
