--- components/ui/Answer/answer.module.css
@@ -1,13 +1,0 @@
-.fadeIn {
-  animation: fadeIn 0.5s ease-in-out forwards;
-  opacity: 0;
-}
-
-@keyframes fadeIn {
-  from {
-    opacity: 0;
-  }
-  to {
-    opacity: 1;
-  }
-}
