+++ components/ui/MemoCard.tsx
