+++ pages/api/create-checkout-session.ts
@@ -43,7 +43,7 @@ const CreateCheckoutSession: NextApiHandler = async (req, res) => {
 
       return res.status(200).json({ sessionId: session.id });
     } catch (err: any) {
-      console.log(err);
+      console.error(err);
       res.status(500).json({ error: { statusCode: 500, message: err.message } });
     }
   } else {
