+++ pages/api/webhooks.ts
@@ -53,7 +53,7 @@ const webhookHandler = async (req: NextApiRequest, res: NextApiResponse) => {
       if (!sig || !webhookSecret) return;
       event = stripe.webhooks.constructEvent(buf, sig, webhookSecret);
     } catch (err: any) {
-      console.log(`❌ Error message: ${err.message}`);
+      console.error(`❌ Error message: ${err.message}`);
       return res.status(400).send(`Webhook Error: ${err.message}`);
     }
 
@@ -89,7 +89,7 @@ const webhookHandler = async (req: NextApiRequest, res: NextApiResponse) => {
             throw new Error("Unhandled relevant event!");
         }
       } catch (error) {
-        console.log(error);
+        console.error(error);
         return res.status(400).send('Webhook error: "Webhook handler failed. View logs."');
       }
     }
