+++ utils/supabase-client.ts
@@ -15,7 +15,7 @@ export const getActiveProductsWithPrices = async (): Promise<ProductWithPrice[]>
     .order("unit_amount", { foreignTable: "prices" });
 
   if (error) {
-    console.log(error.message);
+    console.error(error.message);
   }
   // TODO: improve the typing here.
   return (data as any) || [];
